Added flash address alignment in flash_write_word()...

+++ keepkey_board/public/keepkey_flash.h
@@ -31,7 +31,7 @@
 void flash_erase(Allocation group);
 void flash_erase_word(Allocation group);
 void flash_write(Allocation group, uint32_t offset, uint32_t len, uint8_t* data);
-bool flash_write_word(Allocation group, uint32_t offset, uint32_t len, uint32_t* data);
+bool flash_write_word(Allocation group, uint32_t offset, uint32_t len, uint8_t* data);
 
 
 #endif
